Campsite Search Functionality
So basically this is for the campsite search. I wanted it to look like google where you have a search bar at the center of the screen and after searching, the results appear below. I started by trying to modify the search page itself then when that didn't work I made a new page that displays the results. So when you click "search" the page changes to the new campsite search page with the results displayed. Now I need to format how the data is displayed and such.