Polished Campsite Features and added Mountain Functionality
Few changes. First, I added a "No Results" prompt when the user searches for data (be on the search page or the display page) and nothing matches. It still needs work. Right now its a single column on the tableWidget but I'm having trouble resizing it dynamically so I think I'll change the "No Results" to a pop-up message. Second, I optimized some code by making functions to get rid of some repeating code so the code should look a bit nicer. Third and finally, I added functionality to the search mountain, display mountains, and create mountain pages. For the create mountain pages to work, I add to rework the create_item() function on CampingDatabase_SQLite a little bit but now it all works accordingly.